Renamed everything to snake_case
