+++ Eclipse/src/eclipse/Events/key_events.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Event.h"
+#include "event.h"
 
 namespace eclipse {
