+++ Eclipse/src/eclipse/Layer.cpp
@@ -1,5 +1,5 @@
 #include "ecpch.h"
-#include "Layer.h"
+#include "layer.h"
 
 namespace eclipse {
