+++ Eclipse/src/eclipse/Log.cpp
@@ -1,5 +1,5 @@
 #include "ecpch.h"
-#include "Log.h"
+#include "log.h"
 
 #include "spdlog/sinks/stdout_color_sinks.h"
