+++ Eclipse/src/eclipse/Log.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Core.h"
+#include "core.h"
 #include "spdlog/spdlog.h"
 #include "spdlog/fmt/ostr.h"
