+++ Eclipse/src/eclipse/Window.h
@@ -2,9 +2,9 @@
 
 #include "ecpch.h"
 
-#include "eclipse/Core.h"
-#include "eclipse/Events/Event.h"
-#include "eclipse/CommonTypes/WindowSize.h"
+#include "eclipse/core.h"
+#include "eclipse/events/event.h"
+#include "eclipse/common_types/window_size.h"
 
 namespace eclipse {
