+++ Eclipse/src/eclipse/common_types/mouse_scroll_offset.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "eclipse/Core.h"
+#include "eclipse/core.h"
 
 namespace eclipse {
