+++ Eclipse/src/eclipse/common_types/point_2d.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "eclipse/Core.h"
+#include "eclipse/core.h"
 
 namespace eclipse {
