+++ Eclipse/src/eclipse/imgui/imgui_layer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "eclipse/Layer.h"
+#include "eclipse/layer.h"
 
 namespace eclipse {
