+++ Eclipse/src/platform/opengl/imgui_opengl_renderer.cpp
@@ -55,7 +55,7 @@
 
 #include "ecpch.h"
 #include "imgui.h"
-#include "ImGuiOpenGLRenderer.h"
+#include "imgui_opengl_renderer.h"
 #include "glad/glad.h"
 #include <stdio.h>
 #if defined(_MSC_VER) && _MSC_VER <= 1500 // MSVC 2008 or earlier
