+++ Sandbox/Sandbox.vcxproj
@@ -137,7 +137,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="src\SandboxApp.cpp" />
+    <ClCompile Include="src\sandbox_app.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Eclipse\Eclipse.vcxproj">
